pyproject.toml
[tool.poetry]
name = "github-runner-manager"
version = "0.1.0"
description = "GitHub Runner Manager CLI for Docker-based GitHub Actions Runners"
authors = ["Grégory LEFER <glefer@itroom.fr>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
rich = "^13.0.0"
typer = "^0.17"
requests = "^2.32.5"
docker = "^7.1.0"
flake8 = "^7.3.0"
pydantic = "^2.11.9"
pyyaml = "^6.0.2"
dotenv = "^0.9.9"
pytest-xdist = "^3.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.6.1"
black = "^24.0.0"
isort = "^5.12.0"
mypy = "^1.8.0"
pre-commit = "^3.4.0"
types-pyyaml = "^6.0.12.20250915"
types-docker = "^7.1.0.20250916"
types-requests = "^2.32.4.20250913"

[tool.poetry.scripts]
github-runner-manager = "main:main"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
